[BREAKING] Use a different approach to calculating latency
A ping-pong setup (what was being done before) has the disadvantage of requiring N replies to your 1 event, making federation storms very possible.

The new approach uses a more lazy-loaded approach that keeps the number of events to send at 1. Pings now contain information about the server's processing times so that other servers can report on it. This gets rid of the pong and merges it into the pings.

This is classified as a breaking change because older bots will still pong the pings. They'll see the new bots as not receiving their events as well. They should not crash or break, however. Further, a bunch of metrics were removed becuase they are no longer possible or relevant.